(Claude) Create core domain modules (Phase 0 step 1)
Created three core domain modules to centralize fundamental business
logic that was previously embedded in brain.py.

Files created:
- bt_servant_engine/core/intents.py - IntentType enum (14 intents) and UserIntents model
- bt_servant_engine/core/language.py - Language enum, ResponseLanguage/MessageLanguage/TranslatedPassage models, SUPPORTED_LANGUAGE_MAP constant
- bt_servant_engine/core/agentic.py - AgenticStrengthChoice enum, AgenticStrengthSetting model, ALLOWED_AGENTIC_STRENGTH constant

Files modified:
- brain.py - Import from new modules, remove duplicate definitions, remove unused Enum import, fix import grouping

Extracted from brain.py:
- Lines 649-661: supported_language_map → SUPPORTED_LANGUAGE_MAP
- Lines 705-723: Language enum, ResponseLanguage model
- Lines 739-754: ALLOWED_AGENTIC_STRENGTH, AgenticStrengthChoice, AgenticStrengthSetting
- Lines 801-818: MessageLanguage, TranslatedPassage
- Lines 1120-1139: IntentType, UserIntents

Helper functions (_resolve_agentic_strength, _model_for_agentic_strength)
kept in brain.py for now (they depend on BrainState and config).

Test plan:
- All linters pass (ruff, pylint, mypy, pyright)
- Import linter passes (no architectural violations)
- Tests pass (pytest with coverage ≥70%)

Next: Create 4 helper service modules
